Try to enable timekeeping during oscilloscope use.
Gabriel (@ganzziani), is this the correct way to reenable timekeeping?

Previously, in the main loop timekeeping was disabled upon entering
oscilloscope, and reenabled upon return.

Note that there are sections in the oscilloscope code where timekeeping
is briefly disabled (via "RTC.INTCTRL = 0x00;"). I found a commented-out
line in the oscilloscope code that reenables timekeeping (in a different
mode than expected by this oscilloscope platform; probably the different
mode is from the previous oscilloscope platform, from which this code
was ported). I uncommented this line, switching it to the expected mode.

Then in the main loop, I commented-out the lines that disabled
timekeeping before entering the oscilloscope mode, and that reenabled
timekeeping after leaving the mode.